Add files via upload
Added "success" and "error" messages support to every page, except for board. A small adjustment to mainPage.ejs (adding an empty div, so the messages display properly). Added messages for logging in, logging out, creating a new group, joining a group, removing an user from a group and creating a lesson.